+++ deeppavlov/deep.py
@@ -22,12 +22,12 @@ from deeppavlov.core.commands.train import train_evaluate_model_from_config
 from deeppavlov.core.common.cross_validation import calc_cv_score
 from deeppavlov.core.common.file import find_config
 from deeppavlov.download import deep_download
-from utils.alexa.server import run_alexa_default_agent
-from utils.alice import start_alice_server
-from utils.ms_bot_framework_utils.server import run_ms_bf_default_agent
-from utils.pip_wrapper import install_from_config
-from utils.server_utils.server import start_model_server
-from utils.telegram_utils.telegram_ui import interact_model_by_telegram
+from deeppavlov.utils.alexa.server import run_alexa_default_agent
+from deeppavlov.utils.alice import start_alice_server
+from deeppavlov.utils.ms_bot_framework.server import run_ms_bf_default_agent
+from deeppavlov.utils.pip_wrapper import install_from_config
+from deeppavlov.utils.server.server import start_model_server
+from deeppavlov.utils.telegram.telegram_ui import interact_model_by_telegram
 
 log = getLogger(__name__)
