+++ deeppavlov/utils/ms_bot_framework/conversation.py
