+++ deeppavlov/utils/settings/models_info.json
