+++ docs/devguides/amazon_alexa.rst
@@ -179,7 +179,7 @@ with valid certificate from CA. `Here is the guide <https://developer.amazon.com
 for running custom skill service with self-signed certificates in test mode.
 
 Your intent and slot names defined in Alexa Development Console should be the same with values defined in
-DeepPavlov settings file ``utils/settings/server_config.json``. JSON examples from this guide use default values from
+DeepPavlov settings file ``deeppavlov/utils/settings/server_config.json``. JSON examples from this guide use default values from
 the settings file.
 
 DeepPavlov skill/component can be made available for Amazon Alexa as a REST service by:
@@ -204,5 +204,5 @@ for **each** conversation.
 You should use ``--no-default-skill`` optional flag if your component implements an interface of DeepPavlov *Skill*
 to skip its wrapping with DeepPavlov *DefaultStatelessSkill*.
 
-REST service properties (host, port, https options) are provided in ``utils/settings/server_config.json``. Please note,
+REST service properties (host, port, https options) are provided in ``deeppavlov/utils/settings/server_config.json``. Please note,
 that all command line parameters override corresponding config ones.
